MOEAD_Fixes (#62)
MOEAD now works with the new DESDEO. I also checked an interactive version with references point as the preferences information. Seems to work. More tests should be conducted.

I also added the feature to save non-dominated solutions from each generation. I saw it in the original paper but it was missing from the implementation. The archive is rudimentary, but it works.